+++ src/MQTTPacketOut.c
@@ -148,7 +148,7 @@ int MQTTPacket_send_pingreq(networkHandles* net, char* clientID)
  * @param clientID the string client identifier, only used for tracing
  * @return the completion code (e.g. TCPSOCKET_COMPLETE)
  */
-int MQTTPacket_send_subscribe(List* topics, List* qoss, int msgid, int dup, networkHandles* net, char* clientID)
+int MQTTPacket_send_subscribe(List* topics, List* qoss, uint64_t msgid, int dup, networkHandles* net, char* clientID)
 {
 	Header header;
 	char *data, *ptr;
@@ -220,7 +220,7 @@ void* MQTTPacket_suback(unsigned char aHeader, char* data, int datalen)
  * @param clientID the string client identifier, only used for tracing
  * @return the completion code (e.g. TCPSOCKET_COMPLETE)
  */
-int MQTTPacket_send_unsubscribe(List* topics, int msgid, int dup, networkHandles* net, char* clientID)
+int MQTTPacket_send_unsubscribe(List* topics, uint64_t msgid, int dup, networkHandles* net, char* clientID)
 {
 	Header header;
 	char *data, *ptr;
